# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TuckZone Canteen API",
        "description": "Backend for the school canteen pre-order app.\n\nAuthentication: obtain tokens from /api/auth/login or\n/api/auth/otp/login, then send `Authorization: Bearer <accessToken>`.\nWhen a call returns 401, exchange the refresh token at\n/api/auth/refresh rather than forcing the user to sign in again.\n\nOrdering is idempotent: generate one idempotencyKey when the\ncheckout screen opens and reuse it for every retry of that\nsame order, otherwise a double tap creates two orders.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/devices": {
            "post": {
                "tags": [
                    "notification-controller"
                ],
                "operationId": "registerDevice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeviceTokenRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "notification-controller"
                ],
                "operationId": "unregisterDevice",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/auth/register/teacher": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "registerTeacher",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TeacherRegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register/student": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "registerStudent",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentRegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register/parent": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "registerParent",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParentRegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register/admin": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "registerAdmin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminRegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "refresh",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/password/reset": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "resetPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/auth/otp/request": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "requestOtp",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OtpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OtpIssuedResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/otp/login": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "loginWithOtp",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OtpLoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "logout",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LogoutRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "notification-controller"
                ],
                "operationId": "feed",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NotificationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserSummary": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "STUDENT",
                            "PARENT",
                            "TEACHER",
                            "CANTEEN_ADMIN"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACTIVE",
                            "DISABLED"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "admissionNumber": {
                        "type": "string"
                    },
                    "studentClass": {
                        "type": "string"
                    },
                    "section": {
                        "type": "string"
                    },
                    "rollNumber": {
                        "type": "string"
                    },
                    "seatNumber": {
                        "type": "string"
                    },
                    "studentMobile": {
                        "type": "string"
                    },
                    "parentMobile": {
                        "type": "string"
                    },
                    "employeeId": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    }
                }
            },
...
            "DeviceTokenRequest": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "maxLength": 512,
                        "minLength": 0
                    },
                    "platform": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 0
                    }
                },
                "required": [
                    "platform",
                    "token"
                ]
            },
            "TeacherRegisterRequest": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 180,
                        "minLength": 0
                    },
                    "mobile": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[6-9]\\d{9}$"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 72,
                        "minLength": 8
                    },
                    "employeeId": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 0
                    },
                    "department": {
                        "type": "string",
                        "maxLength": 80,
                        "minLength": 0
                    }
                },
                "required": [
                    "department",
                    "email",
                    "employeeId",
                    "fullName",
                    "mobile",
                    "password"
                ]
            },
            "StudentRegisterRequest": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 180,
                        "minLength": 0
                    },
                    "mobile": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[6-9]\\d{9}$"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 72,
                        "minLength": 8
                    },
                    "admissionNumber": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 0
                    },
                    "studentClass": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 0
                    },
                    "section": {
                        "type": "string",
                        "maxLength": 10,
                        "minLength": 0
                    },
                    "rollNumber": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 0
                    },
                    "seatNumber": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 0
                    },
                    "parentMobile": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[6-9]\\d{9}$"
                    },
                    "studentMobile": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 0
                    }
                },
                "required": [
                    "admissionNumber",
                    "email",
                    "fullName",
                    "mobile",
                    "parentMobile",
                    "password",
                    "rollNumber",
                    "section",
                    "studentClass"
                ]
            },
            "ParentRegisterRequest": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 180,
                        "minLength": 0
                    },
                    "mobile": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[6-9]\\d{9}$"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 72,
                        "minLength": 8
                    }
                },
                "required": [
                    "email",
                    "fullName",
                    "mobile",
                    "password"
                ]
            },
            "AdminRegisterRequest": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 180,
                        "minLength": 0
                    },
                    "mobile": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[6-9]\\d{9}$"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 72,
                        "minLength": 8
                    },
                    "signupCode": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "fullName",
                    "mobile",
                    "password",
                    "signupCode"
                ]
            },
            "RefreshRequest": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "refreshToken"
                ]
            },
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "tokenType": {
                        "type": "string"
                    },
                    "accessToken": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "expiresInSeconds": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserSummary"
                    }
                }
            },
            "PasswordResetRequest": {
                "type": "object",
                "properties": {
                    "mobile": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[6-9]\\d{9}$"
                    },
                    "code": {
                        "type": "string",
                        "minLength": 1
                    },
                    "newPassword": {
                        "type": "string",
                        "maxLength": 72,
                        "minLength": 8
                    }
                },
                "required": [
                    "code",
                    "mobile",
                    "newPassword"
                ]
            },
            "OtpRequest": {
                "type": "object",
                "properties": {
                    "mobile": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[6-9]\\d{9}$"
                    },
                    "purpose": {
                        "type": "string",
                        "enum": [
                            "LOGIN",
                            "PASSWORD_RESET"
                        ]
                    }
                },
                "required": [
                    "mobile",
                    "purpose"
                ]
            },
            "OtpIssuedResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "expiresInMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "devCode": {
                        "type": "string"
                    }
                }
            },
            "OtpLoginRequest": {
                "type": "object",
                "properties": {
                    "mobile": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[6-9]\\d{9}$"
                    },
                    "code": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "code",
                    "mobile"
                ]
            },
            "LogoutRequest": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "refreshToken"
                ]
            },
            "LoginRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
...
            "NotificationResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "event": {
                        "type": "string",
                        "enum": [
                            "ORDER_PLACED",
                            "PAYMENT_SUCCESSFUL",
                            "ORDER_ACCEPTED",
                            "ORDER_PREPARING",
                            "ORDER_PACKED",
                            "ORDER_OUT_FOR_DELIVERY",
                            "ORDER_DELIVERED",
                            "ORDER_CANCELLED",
                            "ORDER_REJECTED",
                            "REFUND_ISSUED",
                            "WALLET_TOPPED_UP",
                            "PASSWORD_CHANGED",
                            "ORDERING_OPENED",
                            "ORDERING_CLOSED"
                        ]
                    },
                    "title": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "payload": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}